Many changes to bring code into parity with current meeloo version

+++ deps/tracemonkey/jshash.h
@@ -49,7 +49,7 @@
 
 JS_BEGIN_EXTERN_C
 
-typedef uint64 JSHashNumber;
+typedef uint32 JSHashNumber;
 typedef struct JSHashEntry JSHashEntry;
 typedef struct JSHashTable JSHashTable;
